requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2
djangorestframework==3.12.4
flake8==3.9.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
maxminddb==2.0.3
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyOpenSSL==20.0.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
zope.interface==5.4.0
